KIS/requirements.txt
# 8. requirements.txt
#
requests==2.26.0
pyyaml==6.0
#